feat: Phase 77: Visual Identity Enhancement
This phase focuses on elevating the application's brand identity by implementing a custom visual logo. Following client feedback, we will replace the generic icon and text-based title on the Home Page with a bespoke SVG logo component for "Yaygara". This new logo will be styled in the "Kid Playful" aesthetic and will feature a smooth "slide up" entrance animation, creating a more cohesive and professional landing experience for the users.